+++ threedi_model_migration/schematisation.py
@@ -1,8 +1,12 @@
+from .file import File
+from .file import Raster
 from dataclasses import dataclass
 from datetime import datetime
 from pathlib import Path
 from typing import List
 from typing import Optional
+from typing import Set
+from typing import Union
 
 
 __all__ = ["Schematisation", "SchemaRevision"]
@@ -20,6 +24,10 @@ class SchemaRevision:
     commit_msg: str
     commit_user: str
 
+    # files
+    sqlite: File
+    rasters: List[Raster]
+
     def __repr__(self):
         return f"SchemaRevision({self.revision_nr})"
 
@@ -35,6 +43,15 @@ class Schematisation:
     @property
     def concat_name(self):
         return f"{self.slug}-{self.sqlite_name}-{self.settings_name}"
+
+    def get_files(self) -> Set[Union[File, Raster]]:
+        result = set()
+        for revision in self.revisions:
+            result.add(revision.sqlite)
+            for raster in revision.rasters:
+                result.add(raster)
+
+        return result
 
     def __repr__(self):
         return f"Schematisation({self.slug}, sqlite_name={self.sqlite_name}, settings_name={self.settings_name})"