+++ threedi_model_migration/tests/test_repository.py
@@ -1,11 +1,21 @@
+from pathlib import Path
+from threedi_model_migration.file import RasterOptions
+
+import pytest
+
+
 def test_revisions(repository_inspected):
     revisions = repository_inspected.revisions
     assert len(revisions) == 2
 
     assert revisions[0].commit_msg == "My second commit"
     assert revisions[0].revision_nr == 2
+    assert len(revisions[0].changes) == 1
+    assert revisions[0].changes[0].path.name == "db2.sqlite"
     assert revisions[1].commit_msg == "My first commit"
     assert revisions[1].revision_nr == 1
+    assert len(revisions[1].changes) == 1
+    assert revisions[1].changes[0].path.name == "db1.sqlite"
 
 
 def test_checkout_newest(repository):
@@ -34,8 +44,19 @@ def test_settings(repository_inspected):
 
     assert settings[0].settings_id == 1
     assert settings[0].settings_name == "default"
+    assert len(settings[0].rasters) == 1
+    assert settings[0].rasters[0].raster_type == RasterOptions.dem_file.value
+    assert settings[0].rasters[0].path == Path("rasters/dem.tif")
     assert settings[1].settings_id == 2
-    assert settings[1].settings_name == "breach"
+    assert settings[1].settings_name == "groundwater"
+    assert len(settings[1].rasters) == 2
+    assert settings[1].rasters[0].raster_type == RasterOptions.dem_file.value
+    assert settings[1].rasters[0].path == Path("rasters/dem.tif")
+    assert (
+        settings[1].rasters[1].raster_type
+        == RasterOptions.groundwater_impervious_layer_level_file.value
+    )
+    assert settings[1].rasters[1].path == Path("rasters/x.tif")
 
 
 def test_inspect(repository_inspected):
@@ -58,3 +79,28 @@ def test_inspect(repository_inspected):
     assert revision.revision_nr == 1
     assert sqlite.sqlite_path.name == "db1.sqlite"
     assert settings.settings_id == 1
+
+
+@pytest.mark.parametrize(
+    "revision_nr,path",
+    [
+        (2, "db1.sqlite"),
+        (2, "db2.sqlite"),
+        (1, "db1.sqlite"),
+    ],
+)
+def test_get_file(revision_nr, path, repository_inspected):
+    file = repository_inspected.get_file(revision_nr, Path(path))
+    assert file.path.name == path
+
+
+@pytest.mark.parametrize(
+    "revision_nr,path",
+    [
+        (1, "db2.sqlite"),
+        (2, "db3.sqlite"),
+    ],
+)
+def test_get_file_not_found(revision_nr, path, repository_inspected):
+    with pytest.raises(FileNotFoundError):
+        repository_inspected.get_file(revision_nr, Path(path))